setting: file structure에 따른 파일 생성 #1
