+++ gomintapa/lib/src/app.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+import 'screens/home.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      routes: {
+        '/': (context) => Home(),
+      },
+      initialRoute: '/',
+      // home: Intro(),
+    );
+  }
+}
